+++ crates/pvm/parser/src/format/rii.rs
@@ -1,3 +1,5 @@
+//! The RII instruction format.
+
 use crate::format::{ISA, RII};
 
 impl From<RII> for Vec<u8> {
@@ -27,8 +29,8 @@ impl From<&[u8]> for RII {
         let mid = (bytes.len() - 1).min(4) + 1;
         RII {
             reg0: bytes[0] % 16,
-            imm0: u32::read_imm(&bytes[1..mid]),
-            imm1: u32::read_imm(&bytes[mid..]),
+            imm0: u64::read_imm(&bytes[1..mid]),
+            imm1: u64::read_imm(&bytes[mid..]),
         }
     }
 }
